Fix paging

+++ src/model/search_result.rs
@@ -4,6 +4,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct SearchResult {
-    pub pages: i32,
+    pub pages: usize,
     pub files: Vec<MarkdownHeader>,
 }
